Refs #1142: Validating collection on PostGres query

+++ Idno/Data/Postgres.php
@@ -142,6 +142,10 @@
                         return $array['_id'];
                     }
                 }*/
+                $collections = $this->getAvailableCollections();
+                if (!in_array($collection, $collections)) {
+                    throw new \Exception("Unsupported collection $collection");
+                }
 
                 if (empty($array['_id'])) {
                     $array['_id'] = md5(rand() . microtime(true));
@@ -271,6 +275,11 @@
             function getRecordByUUID($uuid, $collection = 'entities')
             {
                 try {
+                    $collections = $this->getAvailableCollections();
+                    if (!in_array($collection, $collections)) {
+                        throw new \Exception("Unsupported collection $collection");
+                    }
+                    
                     $statement = $this->client->prepare("select distinct {$collection}.* from " . $collection . " where uuid = :uuid");
                     if ($statement->execute(array(':uuid' => $uuid))) {
                         return $statement->fetch(\PDO::FETCH_ASSOC);
@@ -315,6 +324,11 @@
 
             function getRecord($id, $collection = 'entities')
             {
+                $collections = $this->getAvailableCollections();
+                if (!in_array($collection, $collections)) {
+                    throw new \Exception("Unsupported collection $collection");
+                }
+                
                 $statement = $this->client->prepare("select {$collection}.* from " . $collection . " where _id = :id");
                 if ($statement->execute(array(':id' => $id))) {
                     return $statement->fetch(\PDO::FETCH_ASSOC);
@@ -332,6 +346,11 @@
             function getAnyRecord($collection = 'entities')
             {
                 try {
+                    $collections = $this->getAvailableCollections();
+                    if (!in_array($collection, $collections)) {
+                        throw new \Exception("Unsupported collection $collection");
+                    }
+                
                     $statement = $this->client->prepare("select {$collection}.* from " . $collection . " limit 1");
                     if ($statement->execute()) {
                         if ($row = $statement->fetch(\PDO::FETCH_ASSOC)) {
@@ -438,7 +457,11 @@
             function getRecords($fields, $parameters, $limit, $offset, $collection = 'entities')
             {
                 try {
-
+                    $collections = $this->getAvailableCollections();
+                    if (!in_array($collection, $collections)) {
+                        throw new \Exception("Unsupported collection $collection");
+                    }
+                    
                     // Build query
                     $query            = "select distinct {$collection}.* from {$collection} ";
                     $variables        = array();
@@ -667,6 +690,11 @@
             {
                 try {
 
+                    $collections = $this->getAvailableCollections();
+                    if (!in_array($collection, $collections)) {
+                        throw new \Exception("Unsupported collection $collection");
+                    }
+                    
                     // Build query
                     $query            = "select count(distinct {$collection}.uuid) as total from {$collection} ";
                     $variables        = array();
@@ -719,6 +747,11 @@
             {
                 try {
 
+                    $collections = $this->getAvailableCollections();
+                    if (!in_array($collection, $collections)) {
+                        throw new \Exception("Unsupported collection $collection");
+                    }
+                    
                     $client = $this->client;
                     /* @var \PDO $client */
                     $statement = $client->prepare("delete from {$collection} where _id = :id");
